Reorganize files into modules
