+++ .gitignore
@@ -163,6 +163,7 @@ cython_debug/
 
 # Structure 
 draft/
+assets/
 
 # Database
 *.db
